requirements.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
Authlib==1.6.0
Automat==20.2.0
axial_positional_embedding==0.3.12
Babel==2.8.0
backoff==2.2.1
banks==2.2.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
blobfile==3.0.0
botocore==1.23.34
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.7.14
chardet==4.0.0
charset-normalizer==3.4.1
chromadb==1.0.12
click==8.2.1
cloud-init==24.4.1
cloudpickle==3.1.1
colorama==0.4.4
coloredlogs==15.0.1
comm==0.2.2
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.3.2
cryptography==3.4.8
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.5.1
dbus-python==1.2.18
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
dirtyjson==1.0.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker-pycreds==0.4.0
docstring_parser==0.16
docutils==0.17.1
durationpy==0.10
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.1
filelock==3.13.1
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2024.6.1
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-adk==1.7.0
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.104.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.35.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-texttospeech==2.27.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.26.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.3
griffe==1.7.3
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.20.2
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.4
humanfriendly==10.0
hyper-connections==0.1.15
hyperlink==21.0.0
idna==3.3
importlib_metadata==8.7.0
importlib_resources==6.5.2
incremental==21.3.0
inquirerpy==0.3.4
ipykernel==6.29.5
ipython==8.37.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kagglehub==0.3.12
keyring==23.5.0
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.64
langchain-openai==0.3.21
langchain-text-splitters==0.3.8
langchainhub==0.1.21
langsmith==0.3.45
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lightning==2.5.1.post0
lightning-utilities==0.14.3
litellm==1.74.4
llama-cloud==0.1.32
llama-cloud-services==0.6.43
llama-index==0.12.50
llama-index-agent-openai==0.4.12
llama-index-cli==0.4.4
llama-index-core==0.12.50
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.7.10
llama-index-instrumentation==0.3.0
llama-index-llms-openai==0.4.7
llama-index-multi-modal-llms-openai==0.5.3
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.11
llama-index-readers-llama-parse==0.4.0
llama-index-workflows==1.1.0
llama-parse==0.6.43
local-attention==1.11.1
loguru==0.7.3
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==2.0.1
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mcp==1.11.0
mdurl==0.1.2
mmh3==5.1.0
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==1.48.0
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.3
nltk==3.9.1
numpy==2.1.2
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.7.1.26
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.8.55
oauthlib==3.2.2
olefile==0.46
omegaconf==2.3.0
onnxruntime==1.22.0
openai==1.84.0
opencv-python==4.12.0.88
opentelemetry-api==1.34.0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-exporter-otlp-proto-common==1.34.0
opentelemetry-exporter-otlp-proto-grpc==1.34.0
opentelemetry-instrumentation==0.55b0
opentelemetry-instrumentation-asgi==0.55b0
opentelemetry-instrumentation-fastapi==0.55b0
opentelemetry-proto==1.34.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.34.0
opentelemetry-semantic-conventions==0.55b0
opentelemetry-util-http==0.55b0
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.16.0
pexpect==4.9.0
pfzy==0.3.4
pillow==11.0.0
platformdirs==4.3.7
posthog==4.4.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycryptodomex==3.22.0
pycurl==7.44.1
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pyfaidx==0.8.1.3
Pygments==2.19.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pypdf==5.8.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyrsistent==0.18.1
pyserial==3.5
PySocks==1.7.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytorch-lightning==2.5.1.post0
pytz==2022.1
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
roman==3.3
rpds-py==0.25.1
rsa==4.8
s3transfer==0.5.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
SecretStorage==3.3.1
sentencepiece==0.2.0
sentry-sdk==2.27.0
service-identity==18.1.0
setproctitle==1.3.6
shapely==2.1.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sqlitedict==2.1.0
sse-starlette==2.4.1
stack-data==0.6.3
starlette==0.47.1
streamlit==1.47.0
striprtf==0.0.26
sympy==1.13.3
systemd-python==234
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.7.0+cu128
torchao==0.10.0
torchaudio==2.7.0+cu128
torchdata==0.11.0
torchmetrics==1.7.1
torchtune==0.6.1
torchvision==0.22.0+cu128
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton==3.3.0
Twisted==22.1.0
typer==0.16.0
types-requests==2.32.0.20250602
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
wadllib==1.3.6
wandb==0.19.10
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.23.0
zope.interface==5.4.0
zstandard==0.23.0
